Cleaning up a bit the messages
