+++ common/Cargo.toml
@@ -70,3 +70,4 @@ simplelog = "0.10"
 anyhow = "1"
 flexi_logger = "0.19"
 env_logger = "0.9.0"
+transitive_from = "0.1.3"
