--- common/src/node/logic.rs
@@ -1,3 +1,0 @@
-pub mod messages;
-pub mod stats;
-pub mod text_messages;
