+++ common/src/node/modules/mod.rs
@@ -1,2 +1,4 @@
 pub mod random_connections;
 pub mod gossip_chat;
+pub mod messages;
+mod text_messages_v1;
